+++ src/main/java/congregamystica/integrations/appliedenergistics2/additions/IntegrationsAE2.java
@@ -2,6 +2,8 @@ package congregamystica.integrations.appliedenergistics2.additions;
 
 import appeng.core.Api;
 import congregamystica.api.IAddition;
+import congregamystica.api.IProxy;
+import congregamystica.config.ConfigHandlerCM;
 import congregamystica.utils.helpers.RegistryHelper;
 import congregamystica.utils.libs.ModIds;
 import net.minecraft.block.Block;
@@ -17,27 +19,53 @@ import thaumcraft.api.aspects.AspectList;
 
 import java.util.Map;
 
-public class IntegrationsAE2 implements IAddition {
+public class IntegrationsAE2 implements IAddition, IProxy {
+    private boolean certusClusterEnabled;
+    private boolean chargedClusterEnabled;
+
+    @Override
+    public void preInit() {
+        for(String str : ConfigHandlerCM.clusters.additionalClusters) {
+            if(str.startsWith("oreQuartzCertus;")) {
+                this.certusClusterEnabled = true;
+            } else if(str.startsWith("oreChargedQuartzCertus;")) {
+                this.chargedClusterEnabled = true;
+            }
+        }
+    }
+
     @Override
     public void registerAspects(AspectEventProxy registry, Map<ItemStack, AspectList> aspectMap) {
-        registry.registerObjectTag("crystalCertusQuartz",       new AspectList().add(Aspect.CRYSTAL, 5).add(Aspect.ENERGY, 2));
-        registry.registerObjectTag("gemChargedQuartzCertus",    new AspectList().add(Aspect.CRYSTAL, 5).add(Aspect.ENERGY, 10));
+        if(this.certusClusterEnabled) {
+            registry.registerObjectTag("crystalCertusQuartz", new AspectList().add(Aspect.CRYSTAL, 5).add(Aspect.ENERGY, 2));
+        }
+        if(this.chargedClusterEnabled) {
+            registry.registerObjectTag("gemChargedQuartzCertus", new AspectList().add(Aspect.CRYSTAL, 5).add(Aspect.ENERGY, 10));
+        }
     }
 
     @Override
     public void registerOreDicts() {
         Item ae2Material = RegistryHelper.getRegisteredItem(new ResourceLocation(ModIds.ConstIds.applied_energistics, "material"));
         if(ae2Material != Items.AIR) {
-            OreDictionary.registerOre("gemQuartzCertus",        new ItemStack(ae2Material, 1, 0));
-            OreDictionary.registerOre("gemChargedQuartzCertus", new ItemStack(ae2Material, 1, 1));
+            if(this.certusClusterEnabled) {
+                OreDictionary.registerOre("gemQuartzCertus", new ItemStack(ae2Material, 1, 0));
+            }
+            if(this.chargedClusterEnabled) {
+                OreDictionary.registerOre("gemChargedQuartzCertus", new ItemStack(ae2Material, 1, 1));
+            }
         }
-        Block certusOre = Api.INSTANCE.definitions().blocks().quartzOre().maybeBlock().orElse(Blocks.AIR);
-        if(certusOre != Blocks.AIR) {
-            OreDictionary.registerOre("oreQuartzCertus", certusOre);
+        if(this.certusClusterEnabled) {
+            Block certusOre = Api.INSTANCE.definitions().blocks().quartzOre().maybeBlock().orElse(Blocks.AIR);
+            if (certusOre != Blocks.AIR) {
+                OreDictionary.registerOre("oreQuartzCertus", certusOre);
+            }
         }
-        Block chargedCertusOre = Api.INSTANCE.definitions().blocks().quartzOreCharged().maybeBlock().orElse(Blocks.AIR);
-        if(chargedCertusOre != Blocks.AIR) {
-            OreDictionary.registerOre("oreChargedQuartzCertus", chargedCertusOre);
+        if(this.chargedClusterEnabled) {
+            Block chargedCertusOre = Api.INSTANCE.definitions().blocks().quartzOreCharged().maybeBlock().orElse(Blocks.AIR);
+            if (chargedCertusOre != Blocks.AIR) {
+                OreDictionary.registerOre("oreChargedQuartzCertus", chargedCertusOre);
+            }
         }
     }