+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 ## 1.0.3
 ### Changed
 - Changed particle packet to improve task handling
-- Removed Thaumcraft Fix requirement from Whisperweed
+- Changed Whisperweed so it no longer requires Thaumcraft Fix
+
+### Fixed
+- Fixed Ironwood Golem requiring olive wood
 
 ---